Fill Child Operations after recalculate order
